% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_prediction.R
\name{make_prediction}
\alias{make_prediction}
\title{Get probabilities of surviving on the 1st, 10th, and 20th day in the 74th hunger games}
\usage{
make_prediction(sex, age, career, rating_rand)
}
\arguments{
\item{sex}{integer. 1 = Female, 0 = Male.}

\item{age}{numeric. 12 - 18 years.}

\item{career}{integer. 1 = yes, 0 = no.}

\item{rating_rand}{numeric. Gamekeepers rating. 3- 11.}
}
\value{
a dataframe with probabilities of surving on 1st, 10th, and 20th day.
}
\description{
Get probabilities of surviving on the 1st, 10th, and 20th day in the 74th hunger games
}
\examples{
make_prediction(1, 16, 0, 11)
}
